Add tests for conversion correctness
The script verify.sh reads in pairs of repository HEADS for a git
and a SVN repository (branch/tag names or commit SHA1 for git and
repository paths for SVN). It checks out those heads and runs diff
against the trees.

The *.heads files in this commit contain sensible default lists of
heads for the core modules.